+++ include/libcommute/expression/generator.hpp
@@ -19,6 +19,7 @@
 #include <iostream>
 #include <memory>
 #include <stdexcept>
+#include <string>
 #include <tuple>
 #include <utility>
 
@@ -122,9 +123,17 @@ public:
     f.set(0, this->shared_from_this(), 1);
   }
 
+  // Convert to string
+  virtual std::string to_string() const {
+    std::string s("g^");
+    s += std::to_string(algebra_id()) + "(";
+    s += tuple_to_string(this->indices_) + ")";
+    return s;
+  }
+
   // Stream output
   friend std::ostream& operator<<(std::ostream& os, generator const& g) {
-    return g.print(os);
+    return os << g.to_string();
   }
 
 private:
@@ -139,12 +148,6 @@ protected:
   virtual bool less(generator const& g) const { return indices_ < g.indices_; }
   virtual bool greater(generator const& g) const {
     return indices_ > g.indices_;
-  }
-  // Print to stream
-  virtual std::ostream& print(std::ostream& os) const {
-    os << "g^" << algebra_id() << "(";
-    print_tuple(os, this->indices_);
-    return os << ")";
   }
 };